backend/requirements.txt
Django==5.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

django-cors-headers==4.5.0

pandas==2.2.3
numpy==1.26.4

matplotlib==3.9.3
seaborn==0.13.2

reportlab==4.2.5
Pillow==11.0.0

python-dateutil==2.9.0
pytz==2024.2

